Removed authController.js and validateAccessToken.js files and moved their logic to AuthService.js file
